+++ optillm/plugins/spl/__init__.py
@@ -1,0 +1,3 @@
+"""
+System Prompt Learning (SPL) plugin module initialization.
+"""
