+++ optillm/plugins/spl/data/metrics.json
@@ -1,0 +1,10 @@
+{
+  "total_queries": 1,
+  "strategy_applications": 0,
+  "strategies_created": 0,
+  "strategies_refined": 0,
+  "successful_resolutions": 0,
+  "last_strategy_id": 0,
+  "reasoning_examples_collected": 0,
+  "strategies_merged": 0
+}
